+++ modules/@apostrophecms/doc-type/ui/apos/components/AposDocEditor.vue
@@ -38,6 +38,7 @@
                 :current-fields="currentFields"
                 :trigger-validation="triggerValidation"
                 :utility-rail="false"
+                :url="doc.data._url"
                 v-model="docOtherFields"
               />
             </div>
